Conflicts:
	vowpalwabbit/loss_functions.h

+++ vowpalwabbit/search_graph.cc
@@ -230,7 +230,7 @@ void takedown(Search::search& sch, vector<example*>& /*ec*/)
 
 void add_edge_features_group_fn(task_data&D, float fv, uint32_t fx)
 { example*node = D.cur_node;
-  uint32_t fx2 = fx / D.multiplier;
+  uint32_t fx2 = fx / (uint32_t)D.multiplier;
   for (size_t k=0; k<D.numN; k++)
   { if (D.neighbor_predictions[k] == 0.) continue;
     float fv2 = fv * D.neighbor_predictions[k];
@@ -243,7 +243,7 @@ void add_edge_features_group_fn(task_data&D, float fv, uint32_t fx)
 
 void add_edge_features_single_fn(task_data&D, float fv, uint32_t fx)
 { example*node = D.cur_node;
-  uint32_t fx2 = fx / D.multiplier;
+  uint32_t fx2 = fx / (uint32_t)D.multiplier;
   size_t k = (size_t) D.neighbor_predictions[0];
   feature f = { fv, (uint32_t)(( fx2 + 348919043 * k ) * D.multiplier) & (uint32_t)D.mask };
   node->atomics[neighbor_namespace].push_back(f);
@@ -278,7 +278,7 @@ void add_edge_features(Search::search&sch, task_data&D, uint32_t n, vector<examp
         size_t other_side = (D.directed && (n_in_sink != m_in_sink)) ? (D.K+1) : 0;
         D.neighbor_predictions[ D.pred[m]-1 + other_side ] += 1.;
         pred_total += 1.;
-        last_pred = D.pred[m]-1 + other_side;
+        last_pred = (uint32_t)D.pred[m]-1 + (uint32_t)other_side;
       }
     }
     else
@@ -336,7 +336,7 @@ float macro_f(task_data& D)
       predC += (float)D.confusion_matrix[ IDX(j,k) ];
     }
     if (trueC == 0) continue;
-    float correctC = D.confusion_matrix[ IDX(k,k) ];
+    float correctC = (float)D.confusion_matrix[ IDX(k,k) ];
     count_f1++;
     if (correctC > 0)
     { float pre = correctC / predC;
@@ -349,7 +349,7 @@ float macro_f(task_data& D)
 
 void run(Search::search& sch, vector<example*>& ec)
 { task_data& D = *sch.get_task_data<task_data>();
-  float loss_val = 0.5 / (float)D.num_loops;
+  float loss_val = 0.5f / (float)D.num_loops;
   for (size_t n=0; n<D.N; n++) D.pred[n] = D.K+1;
 
   for (size_t loop=0; loop<D.num_loops; loop++)
@@ -367,7 +367,7 @@ void run(Search::search& sch, vector<example*>& ec)
       Search::predictor P = Search::predictor(sch, n+1);
       P.set_input(*ec[n]);
       if (false && (k > 0))
-      { float min_count = 1e12;
+      { float min_count = 1e12f;
         for (size_t k2=1; k2<=D.K; k2++)
           min_count = min(min_count, D.true_counts[k2]);
         float w = min_count / D.true_counts[k];
@@ -393,7 +393,7 @@ void run(Search::search& sch, vector<example*>& ec)
       // make the prediction
       D.pred[n] = P.predict();
       if (ec[n]->l.cs.costs.size() > 0) // for test examples
-        sch.loss((ec[n]->l.cs.costs[0].class_index == D.pred[n]) ? 0. : (last_loop ? 0.5 : loss_val));
+        sch.loss((ec[n]->l.cs.costs[0].class_index == D.pred[n]) ? 0.f : (last_loop ? 0.5f : loss_val));
 
       if (add_features) del_edge_features(D, n, ec);
     }
@@ -401,7 +401,7 @@ void run(Search::search& sch, vector<example*>& ec)
 
   for (uint32_t n=0; n<D.N; n++)
     D.confusion_matrix[ IDX( ec[n]->l.cs.costs[0].class_index, D.pred[n] ) ] ++;
-  sch.loss( 1. - macro_f(D) );
+  sch.loss( 1.f - macro_f(D) );
 
   if (sch.output().good())
     for (uint32_t n=0; n<D.N; n++)
